- Added memory manager placeholder - Fixed memmap from bootloader - Added ELF section information from bootloader

+++ kernel/src/libk/math.cc
@@ -2,7 +2,7 @@
 
 namespace math {
 
-    long powl(long base, unsigned long exp) {
+    long long powl(long long base, unsigned long long exp) {
         long res = 1;
 
         while (exp) {
@@ -13,9 +13,9 @@ namespace math {
         return res;
     }
 
-    unsigned long logl(unsigned long base, unsigned long arg) {
-        long res                   = 0;
-        unsigned long long temp    = 1;
+    unsigned long long logl(unsigned long long base, unsigned long long arg) {
+        long long res = 0;
+        unsigned long long temp = 1;
         unsigned long long oldTemp = 1;
 
         while ((temp *= base) <= arg) {
